+++ gaia/__init__.py
@@ -1,5 +1,2 @@
 #: Global version number for the package
 __version__ = '0.0.1a1'
-
-from . import inputs
-from . import processes
