--- gaia/inputs/__init__.py
@@ -1,3 +1,0 @@
-from . import formats
-from . import datatypes
-from base import *
